refactor: rename serialized rule
