+++ crates/config/src/rule.rs
